ProxyingRecorder block_digest optimisation
Instead of doing `self.block_digest.update(hunk)` every time we read
some data from the remote server, we do it in the end of the
transmission, when `ProxyingRecorder.close()` is invoked.

I noticed that we update `block_digest` on every remote socket read but
it is not used until much later, when we write to the WARC file.
https://github.com/internetarchive/warcprox/blob/master/warcprox/warc.py#L131

So, we shouldn't block reading with this but we can do in the end of the
transaction and much faster using 1MB chunks instead of 64KB chunks.
The result is exactly the same.

Initial profiling info:
```
ncalls  tottime  percall  cumtime  percall filename:lineno(function)
151718  189.786    0.001  189.786    0.001 {method 'update' of '_hashlib.HASH' objects}
```
After the improvement:
```
12111   74.139    0.006   74.139    0.006 {method 'update' of '_hashlib.HASH' objects}
```