Dedup only urls with Warcprox-Meta and warc-prefix
Warcprox handles requests with HTTP header `Warcprox-Meta` containing
`warc-prefix` in a special way, creating distinct WARC files for them.
In production, we care only for content in these WARCs and discard other
WARCs that use the default `WARCPROX-*` filename.

Thus, I suggest to dedup only responses which have `Warcprox-Meta` and
`warc-prefix`.

A relevant idea is to have an extra flag to enable this behavior
(e.g. `--dedup-only-prefixed`). I'm not sure if this makes sense.