Avoid exception when trying to send error to client
When a request fails for any reason, we use
`MitmProxyHandler.send_error` to send an error response to the client.

The problem is that in many cases, there isn't an open socket connection
With the client any more, so we get a `BrokenPipeError`
`send_error(%r, %r, %r) raised exception`.

We just add a check to make sure that the socket connection with the client is
not None before trying to send the error.