Cache digest_str result in memory to improve performance
We use `warcprox.digest_str` in 2 places during the course of a single
HTTP request. 1) In all `warcprox.dedup` methods to get the key and lookup for
duplicates, 2) in `warcprox.warc` to produce the WARC record.

We use `lru_cache` to avoid recalculating it.

We also reuse the cached result if the request for the same URL is done again.