Increase IO buffer size to improve WarcWriter performance
When we open a WARC file for writing with the standard `open(filename)` we
use a `BufferedWriter` to improve performance. This is the default python
behavior.

`BufferedWriter` uses `io.DEFAULT_BUFFER_SIZE=8192` by default unless we
set a custom value.

We set `io.DEFAULT_BUFFER_SIZE=1048576` (1 MB) to speed up file writing.
The buffer will be written out to the underlying `RawIOBase` object under
various conditions, including:

* when the buffer gets too small for all pending data;
* when flush() is called;
* when the BufferedWriter object is closed or destroyed.

https://docs.python.org/3/library/io.html#io.BufferedWriter

We remove the `flush()` command in the WARC writer.