ssl: add option to enable legacy renegotiation
By default, the ssl library disables insecure renegotiation. However,
there are still servers running the old renegotiation strategy and
these can't be connected to without telling OpenSSL to accept it.
I've tested with one of these servers I've come across live on the
web and I've confirmed that warcprox is unable to capture from it
without this option, but that web browsers (such as Safari, Chrome
and Firefox) are happy to connect to it.

I've made this an option since it's a security option, and made it
opt-in.

refs https://datatracker.ietf.org/doc/html/rfc5746
refs CVE-2009-3555