warcprox: OP_IGNORE_UNEXPECTED_EOF in server too
Unexpected EOFs were previously ignored when sending requests in
be41954460927c43d7dc102d5d05c3acbbf1bd28, but not in the context
used in the HTTP server where we're receiving requests. It looks
like these happen in that context too, so probably best to set
the flag here as well.